+++ search.hpp
@@ -4,9 +4,9 @@
 
 namespace search
 {
-    std::pair<std::vector<double>, double> dijkstra(knn_graph &g, int st, int target);
-    std::pair<std::vector<double>, double> best_first(knn_graph &g, int st, int target);
-    std::pair<std::vector<double>, double> a_algorithim(knn_graph &g, int st);
-    std::pair<std::vector<double>, double> astar(knn_graph &g, int st, int target);
-    std::pair<std::vector<double>, double> dfs(knn_graph &g, int st, int target);
+    std::pair<std::vector<int>, double> dijkstra(knn_graph &g, int st, int target);
+    std::pair<std::vector<int>, double> best_first(knn_graph &g, int st, int target);
+    std::pair<std::vector<int>, double> a_algorithim(knn_graph &g, int st);
+    std::pair<std::vector<int>, double> astar(knn_graph &g, int st, int target);
+    std::pair<std::vector<int>, double> dfs(knn_graph &g, int st, int target);
 }
